+++ mod/wikirate_csv_import/set/right/import_map/core.haml
@@ -12,4 +12,4 @@
     .light-grey-color-2
       (# unmapped)
 
-  = render_tabs
+  = render_map_form
