+++ mod/wikirate_csv_import/set/right/import_map/map_ui.haml
@@ -1,1 +1,1 @@
-%input{ name: "mapping_value" }
+%input{ name: "mapping[#{type}][#{name_in_file}" }
