+++ mod/wikirate_csv_import/set/type/source_import_file.rb
@@ -3,5 +3,5 @@ include_set Abstract::WikirateImport
 attachment :source_import_file, uploader: CarrierWave::FileCardUploader
 
 def import_item_class
-  ImportItem::Structure::SourceCsv
+  SourceImportItem
 end
