+++ mod/wikirate_csv_import/spec/support/shared_table_row_examples.rb
@@ -1,5 +1,5 @@
 shared_context "table_row" do |type_id|
-  def csv_row data
+  def import_item data
     csv_data.merge data
   end
 
@@ -8,8 +8,8 @@ shared_context "table_row" do |type_id|
     io = StringIO.new row_args.values.join(",")
     file = CsvFile.new io, AnswerImportItem
     vm = ValidationManager.new file, :skip
-    vm.validate do |csv_row|
-      yield described_class.new(csv_row, format).render
+    vm.validate do |import_item|
+      yield described_class.new(import_item, format).render
     end
   end
 
@@ -29,8 +29,8 @@ shared_context "table_row" do |type_id|
     io = StringIO.new row_args.values.join(",")
     file = CsvFile.new io, AnswerImportItem
     vm = ValidationManager.new file, :skip
-    vm.validate do |csv_row|
-      @row = described_class.new(csv_row, format).render
+    vm.validate do |import_item|
+      @row = described_class.new(import_item, format).render
       yield
     end
   end
