+++ script/csv_import/metric/import_chrb_metrics.rb
@@ -1,5 +1,5 @@
 require_relative "../../../config/environment"
-require_relative "metric_csv_row"
+require_relative "metric_import_item"
 require_relative "../../../vendor/card-mods/csv_import/lib/import_manager/script_import_manager.rb"
 
 csv_path = File.expand_path "../data/CHRB_import.csv", __FILE__
