+++ script/csv_import/metric/import_hesa_metrics.rb
@@ -1,5 +1,5 @@
 require_relative "../../../config/environment"
-require_relative "metric_csv_row"
+require_relative "metric_import_item"
 require_relative "../csv_file"
 
 csv_path = File.expand_path "../data/hesa_import.csv", __FILE__
