+++ script/csv_import/open_corporates/import_ground_truth.rb
@@ -1,5 +1,5 @@
 require_relative "../../../config/environment"
-require_relative "open_corporates_csv_row"
+require_relative "open_corporates_import_item"
 require_relative "../csv_file"
 
 csv_path = File.expand_path "../data/ground_truth_full_dataset.csv", __FILE__
