+++ script/csv_import/open_corporates/import_headquarters.rb
@@ -1,5 +1,5 @@
 require_relative "../../../config/environment"
-require_relative "headquarters_csv_row"
+require_relative "headquarters_import_item"
 require_relative "../../../mod/csv_import/lib/import_manager/script_import_manager.rb"
 
 csv_path = File.expand_path "../data/additional_headquarter_codes.csv", __FILE__
