+++ script/csv_import/open_corporates/import_oc_mappings_vol1.rb
@@ -1,5 +1,5 @@
 require_relative "../../../config/environment"
-require_relative "open_corporates_csv_row_compact"
+require_relative "open_corporates_import_item_compact"
 
 csv_path = File.expand_path "../data/oc_mappings_vol1.csv", __FILE__
