+++ script/csv_import/open_corporates/import_oc_mappings_vol2.rb
@@ -1,5 +1,5 @@
 require_relative "../../../config/environment"
-require_relative "open_corporates_csv_row_only_headquarters"
+require_relative "open_corporates_import_item_only_headquarters"
 require_relative "../../../mod/csv_import/lib/import_manager/script_import_manager.rb"
 
 csv_path = File.expand_path "../data/oc_mappings_vol2.csv", __FILE__
