+++ script/csv_import/open_corporates/open_corporates_import_item_only_headquarters.rb
@@ -1,0 +1,11 @@
+require_relative "open_corporates_import_item"
+
+# CSV import without incorporation place
+class OpenCorporatesImportItemOnlyHeadquarters < OpenCorporatesImportItem
+  @columns = [:oc_jurisdiction_code, :oc_company_number, :wikirate_number, :company_name]
+  @required = [:oc_jurisdiction_code, :oc_company_number, :wikirate_number]
+
+  def inc_jurisdiction_code
+    nil
+  end
+end
