--- script/csv_import/wikipedia/logo_csv_row.rb
@@ -1,16 +1,0 @@
-require_relative "wikipedia_csv_row"
-
-class LogoImportItem < WikipediaImportItem
-  @columns = [:wikirate_id, :wikirate_name, :wikipedia_url, :logo]
-
-  def normalize_logo val
-    "https:" + val
-  end
-
-  def import
-    return if Card[wikirate_id, :image]
-    puts wikirate_name
-    ensure_card [wikirate_id, :image], remote_image_url: logo,
-                                       type_id: Card::ImageID
-  end
-end
